Added guards, updated controllers and authorization

+++ controllers/delete.js
@@ -2,6 +2,11 @@ module.exports = {
     async get(req, res){
         const id = req.params.id;
         const car = await req.storage.getById(id);
+
+        if(car.owner != req.session.user.id){
+            console.log('User is not owner');
+            return res.redirect('/login');
+        }
         if(car){
             res.render('delete', { title: `Delete - ${car.name}`, car});
         }else{
@@ -10,6 +15,13 @@ module.exports = {
     },
     async post(req, res){
         const id = req.params.id;
+        const car = await req.storage.getById(id);
+
+        if(car.owner != req.session.user.id){
+            console.log('User is not owner');
+            return res.redirect('/login');
+        }
+
         try{
             await req.storage.deleteById(id);
             res.redirect('/');
